iss1 (#2)
* Code styling

* More restructuring
