+++ client/actions/actions.js
@@ -1,12 +1,8 @@
-
 export const FETCH_PULSE = 'FETCH_PULSE'
 
-
-
-
 export function fetchPulse(pulseCount){	
-	return {	
-		type: FETCH_PULSE,
-		payload : pulseCount
-	}
+  return {	
+    type: FETCH_PULSE,
+    payload : pulseCount
+  }
 }
