+++ client/public/index.html
@@ -2,13 +2,9 @@
 <html>
   <head>
     <link rel="stylesheet" href="https://cdn.rawgit.com/twbs/bootstrap/v4-dev/dist/css/bootstrap.css">
-    <style>
-
-    </style>
   </head>
   <body>
     <div id="root"></div>
-    
   </body>
   <script src="../../bundle.js"></script>
 </html>
