--- index.js
@@ -1,16 +1,0 @@
-const express = require('express')
-const http = require('http')
-const bodyParser = require('body-parser')
-const webpack = require('webpack')
-const webpackDevMiddleware = require('webpack-dev-middleware')
-const webpackConfig = require('./webpack.config.js')
-
-
-var app = express();
-
-app.use(express.static(__dirname + '/public'))
-app.use(webpackDevMiddleware(webpack(webpackConfig)))
-app.use(bodyParser.json())
-
-
-app.listen(3000);
