+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "./node_modules/.bin/nodemon index.js --ignore client/"
+    "start": "./node_modules/.bin/nodemon ./server/index.js --ignore client/"
   },
   "author": "",
   "license": "ISC",
