--- public/src/reducers/reducer_pulses.js
@@ -1,13 +1,0 @@
-import {FETCH_PULSE} from '../actions/actions'
-
-
-
-export default function(state=0,action){
-	console.log("payload",action)
-	switch(action.type){
-		case FETCH_PULSE:  
-			return action.payload;  
-
-	}
-	return state;
-}
